OpenGL: Improve accuracy of quaternion interpolation
The max error of doing nlerp on a quaternion vs. slerp is proportional
to the angle of the arc being interpolated across. Therefore, by
reducing the angle we can greatly decrease error. Taking the square root
of a quaternion with a rotation of theta will produce one with a
rotation of theta/2 on the same plane. We do this in the vertex shader
and then reverse the transformation in the fragment shader by squaring.